+++ react-library/__fixtures__/compliant.js
@@ -1,0 +1,7 @@
+export const func = () => {
+  const value = "some value";
+
+  return value;
+};
+
+export default "default export";