+++ react-library/__fixtures__/non-compliant.ts
@@ -1,0 +1,9 @@
+const func = () => {
+  let value = "some value";
+
+  return value;
+};
+
+export function exportedFunc() {
+  return false;
+}