+++ react-library/__fixtures__/non-compliant.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+interface Props {
+  test: string;
+}
+
+const ArrowFunctionComponent: React.FunctionComponent<Props> = (
+  props
+): JSX.Element => <div>{props.test}</div>;