+++ react-library/index.js
@@ -1,0 +1,4 @@
+/* eslint-env node */
+const config = require("./.eslintrc");
+
+module.exports = config;