+++ react/__fixtures__/non-compliant.js
@@ -1,0 +1,5 @@
+const func = () => {
+  let value = "some value";
+
+  return value;
+};